MAINT: update maximum numpy version to build against
The rule is <N+3, and numpy 2.2.x is out, so the bound should
be <2.5.

Also add a `pytest.ini` file that will turn deprecation warnings
in the test suite into errors. This will catch any deprecations
from `numpy` or `python`, so that they can be fixed in time.
Having the test suite run deprecation-warning-free is the guarantee
that the <N+3 upper bound will work.