CI: stop building i686 wheels
NumPy hasn't shipped i686 wheels in a while. This causes the
i686 wheels in this repo to build NumPy from source, and hence
the `ubuntu-latest` wheel build job takes a very long time.
It doesn't make too much sense to ship wheels for a platform
that NumPy doesn't support, so disable `manylinux_i686` and
`musllinux_i686`.